Finalização do projeto

+++ GalloTube/Models/VideoTag.cs
@@ -1,7 +1,7 @@
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
-
 namespace GalloTube.Models;
+
 [Table("VideoTag")]
 public class VideoTag
 {
@@ -11,7 +11,7 @@ public class VideoTag
     public Video Video { get; set; }
 
     [Key, Column(Order = 2)]
-    public byte TagId { get; set; }
+    public int TagId { get; set; }
     [ForeignKey("TagId")]
     public Tag Tag { get; set; }
 }
